modify warnings

+++ sagittarius-common/src/main/java/org/sagittarius/common/File/FileUtil.java
@@ -6,13 +6,21 @@ public class FileUtil {
 
 
     public static void deleteFile(File file) {
-        if (file.isDirectory()) {
-            File[] files = file.listFiles();
-            for (File f : files) {
-                deleteFile(f);
+        if (file.exists()) {
+            if (file.isDirectory()) {
+                File[] files = file.listFiles();
+                if (files != null) {
+                    for (File f : files) {
+                        deleteFile(f);
+                    }
+                }
+            }
+            if (file.isFile()) {
+                if (!file.delete()) {
+                    throw new RuntimeException("Delete file fail!");
+                }
             }
         }
-        file.delete();
-    }
 
+    }
 }
